refactor: update Clickhouse wrapper imports and add new RealClickhouseWrapper for integration testing
- Replaced TestClickhouseWrapper with RealClickhouseWrapper in jest setup and tests.
- Introduced IClickhouseWrapper interface for consistent wrapper implementation.
- Added comprehensive tests for RealClickhouseWrapper to ensure functionality and security.
- Removed outdated MockClickhouseClientWrapper tests and files.
